AddClasses: Fragment -> List
I changed the classes from fragments to the list similar to the notes. It is fairly confusing. I think that there will be a lot of database issues in it. Currently the activity crashes when you try to add a class. The goal is to have a list of classes. The class can be clicked on, opening a Dialog. If you click delete, the course is removed from database.